Lint all files
