+++ src/thor/observations/__init__.py
@@ -1,3 +1,3 @@
-# flake8: noqa: F401
+# ruff: noqa: F401, F403
 from .observations import *
 from .photometry import *
