+++ src/thor/utils/linkages.py
@@ -1,4 +1,3 @@
-import gc
 from typing import Tuple
 
 import numpy as np
